perbaikan tampilan register dan login
