+++ app/screen/index.jsx
@@ -1,21 +1,11 @@
 import HomeScreen from "./home"
 import LoginScreen from "./login"
 import SplashScreen from "./splash"
-import { 
-    RegisterInputNameScreen,
-    RegisterInputDateScreen,
-    RegisterInputEmailScreen,
-    RegisterInputGenderScreen,
-    RegisterInputPasswordScreen
-} from "./register"
+import Register from "./register"
 
 export {
     HomeScreen,
     LoginScreen,
     SplashScreen,
-    RegisterInputNameScreen,
-    RegisterInputDateScreen,
-    RegisterInputEmailScreen,
-    RegisterInputGenderScreen,
-    RegisterInputPasswordScreen
+    Register
 }
