--- app/screen/register/input-date.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function RegisterInputDateScreen(){
-    return(
-        <View>
-            <Text>input date</Text>
-        </View>
-    )
-}
