--- app/screen/register/input-email.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function RegisterInputEmailScreen(){
-    return(
-        <View>
-            <Text>input email</Text>
-        </View>
-    )
-}
