--- app/screen/register/input-gender.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function RegisterInputGenderScreen(){
-    return(
-        <View>
-            <Text>input genderd</Text>
-        </View>
-    )
-}
