--- app/screen/register/input-password.jsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function RegisterInputPasswordScreen(){
-    return(
-        <View>
-            <Text>input password</Text>
-        </View>
-    )
-}
