+++ assets/index.jsx
@@ -1,17 +1,15 @@
-import ICGoogle from './images/icons_google.png'
-import ICFacebook from './images/Frame.png'
 import ICFb from './images/Fb.png'
 import ICLi from './images/LinkedIn.png'
 import ICGgl from './images/Google.png'
 import ICSpsc from './images/splash-sc.png'
+import ICLrs from './images/logres.png'
 
 
 
 export {
-    ICGoogle,
-    ICFacebook,
     ICFb,
     ICLi,
     ICGgl,
-    ICSpsc
+    ICSpsc,
+    ICLrs
 }
